requirements.txt
Django==3.2.4
gunicorn==20.1.0
# psycopg2-binary==2.9.1
django-environ==0.7.0
msrest==0.7.1
pillow==10.2.0
azure-cognitiveservices-vision-computervision==0.9.0

